A_Laboratorios/A_Lab_Intro_to_Airflow/requirements.txt
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
apache-airflow==3.0.0
apache-airflow-providers-standard==1.9.0
python-dotenv>=1.0.0
apache-airflow-providers-databricks==7.7.3
